Upgrades LSTM to take consistent random input
I changed the "prefix" function of LSTM_arg.ipynb to
generate something like what I mentioned in 1/23 update of
the README. I'm not sure if it actually works. That will
be up to future Matthew to figure out.

I was also trying to test out THROCCING, but I ran out of
memory around 700 epochs :(

As a future direction it would be nice to upgrade the Vanilla
LSTM into a stacked to see if it can pick up on cooler patterns.
Maybe lets play around with FRAG_SIZE parameter.