Manipulating Stacked LSTM and amount of data
It seems that it is very difficult not to overfit. This visually
adds examples of what overfitting looks like! The model develops
a decent looking "average" of the training data, and when given
new data, it just moves it around and performs poorly!

It may be useful to, in the future, come back and try different
methods to counter the effects of overfitting. I tried dropout,
but it didn't seem to work very well. It would be useful to learn
how to split up time series data into training and validation sets.